Update CHANGELOG.md; Bump version to v0.1.5; Update all dependencies; Rename internal directory to lib; Fix fieldalignment in a few structs; Remove old Win32 makefile
