+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.5] - TBA
+### Added
+
+### Changed
+- Merged the Windows makefile (`Makefile.win32`) into the default makefile
+- Minimum required version of `make` is v3.82
+
+### Removed
+
 ## [0.1.4] - 2022-02-21
 ### Added
 - Add `go vet` to lint commands
