--- Makefile.win32
@@ -1,19 +1,0 @@
-# Makefile.win32
-#
-lint:
-    go vet ./...
-	golangci-lint run
-
-build:
-	SET CGO_ENABLED=0
-	go build -ldflags "-s -w" -o bin\gowait.exe .\cmd\gowait
-	upx -q bin\gowait.exe
-
-clean:
-	del bin\gowait.exe
-
-build-docker: lint
-	docker build --no-cache -t neflyte/gowait:latest .
-
-bootstrap:
-	scoop install upx
