+++ go.mod
@@ -3,14 +3,9 @@ module github.com/neflyte/gowait
 go 1.16
 
 require (
-	github.com/Shopify/sarama v1.31.1
-	github.com/klauspost/compress v1.14.3 // indirect
-	github.com/lib/pq v1.10.4
-	github.com/neflyte/configmap v0.2.0
-	github.com/sirupsen/logrus v1.8.1
-	golang.org/x/crypto v0.0.0-20220214200702-86341886e292 // indirect
-	golang.org/x/sys v0.0.0-20220209214540-3681064d5158 // indirect
-	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
-	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+	github.com/IBM/sarama v1.42.1
+	github.com/lib/pq v1.10.9
+	github.com/neflyte/configmap v0.3.0
+	github.com/sirupsen/logrus v1.9.3
+	gopkg.in/yaml.v3 v3.0.1
 )
