+++ lib/logger/delegate.go
